A demo of adapting descriptors depending on the context
